Rename /async directory to async_support
`async` is a reserved python keyword since the 3.7 version and thus
cannot be used by a library. I renamed the directory and corrected the
imports accordingly.
