+++ pyrlang/async_support/__init__.py
@@ -13,3 +13,7 @@
 # limitations under the License.
 
 # Compatibility adapters for Gevent and for Python3 asyncio
+from .asyncio_engine import AsyncioEngine
+from .gevent_engine import  GeventEngine
+
+__all__ = ('AsyncioEngine', 'GeventEngine')
